Add files via upload
Image Caption Generation
This project generates captions for images using the Flickr8k dataset. It employs a deep learning model that combines Convolutional Neural Networks (CNNs) for image feature extraction and Recurrent Neural Networks (RNNs) for generating captions.

The goal of the project is to automatically describe the content of images in natural language, which is useful in applications like image search, accessibility, and more.

Dataset
The project uses the Flickr8k dataset, which contains 8,000 images and their corresponding captions. These captions provide descriptive sentences for each image, which the model learns to generate based on the visual content.